@@ -7,7 +7,7 @@ import me.alsturm.reminder.entity.UserSettings;
 import me.alsturm.reminder.repository.UserSettingsRepository;
 import org.springframework.stereotype.Service;
 
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 import java.time.Duration;
 import java.util.Optional;
 import java.util.function.Consumer;
